move FileChange to shared module
